% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.quicksight_operations.R
\name{create_group}
\alias{create_group}
\title{Creates an Amazon QuickSight group}
\usage{
create_group(GroupName, Description = NULL, AwsAccountId, Namespace)
}
\arguments{
\item{GroupName}{[required] A name for the group that you want to create.}

\item{Description}{A description for the group that you want to create.}

\item{AwsAccountId}{[required] The ID for the AWS account that the group is in. Currently, you use the ID for the AWS account that contains your Amazon QuickSight account.}

\item{Namespace}{[required] The namespace. Currently, you should set this to \code{default}.}
}
\description{
Creates an Amazon QuickSight group.
}
\details{
The permissions resource is \code{arn:aws:quicksight:us-east-1:<relevant-aws-account-id>:group/default/<group-name> <relevant-aws-account-id>:group/default/<group-name> }.

The response is a group object.

\strong{CLI Sample:}

\code{aws quicksight create-group --aws-account-id=111122223333 --namespace=default --group-name="Sales-Management" --description="Sales Management - Forecasting" }
}
\section{Accepted Parameters}{
\preformatted{create_group(
  GroupName = "string",
  Description = "string",
  AwsAccountId = "string",
  Namespace = "string"
)
}
}

