% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.quicksight_operations.R
\name{describe_group}
\alias{describe_group}
\title{Returns an Amazon QuickSight group's description and Amazon Resource Name (ARN)}
\usage{
describe_group(
  GroupName = "string",
  AwsAccountId = "string",
  Namespace = "string"
)
}
\arguments{
\item{GroupName}{[required] The name of the group that you want to describe.}

\item{AwsAccountId}{[required] The ID for the AWS account that the group is in. Currently, you use the ID for the AWS account that contains your Amazon QuickSight account.}

\item{Namespace}{[required] The namespace. Currently, you should set this to \code{default}.}
}
\description{
Returns an Amazon QuickSight group's description and Amazon Resource Name (ARN).
}
\details{
The permissions resource is \code{arn:aws:quicksight:us-east-1:<relevant-aws-account-id>:group/default/<group-name> <relevant-aws-account-id>:group/default/<group-name> }.

The response is the group object.

\strong{CLI Sample:}

\code{aws quicksight describe-group -\-aws-account-id=11112222333 -\-namespace=default -\-group-name=Sales }
}
\examples{

}
