% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.quicksight_operations.R
\name{list_group_memberships}
\alias{list_group_memberships}
\title{Lists member users in a group}
\usage{
list_group_memberships(GroupName, NextToken = NULL, MaxResults = NULL,
  AwsAccountId, Namespace)
}
\arguments{
\item{GroupName}{[required] The name of the group that you want to see a membership list of.}

\item{NextToken}{A pagination token that can be used in a subsequent request.}

\item{MaxResults}{The maximum number of results to return from this request.}

\item{AwsAccountId}{[required] The ID for the AWS account that the group is in. Currently, you use the ID for the AWS account that contains your Amazon QuickSight account.}

\item{Namespace}{[required] The namespace. Currently, you should set this to \code{default}.}
}
\description{
Lists member users in a group.
}
\details{
The permissions resource is \code{arn:aws:quicksight:us-east-1:<aws-account-id>:group/default/<group-name> <aws-account-id>:group/default/<group-name> }.

The response is a list of group member objects.

\strong{CLI Sample:}

\code{aws quicksight list-group-memberships -\-aws-account-id=111122223333 -\-namespace=default }
}
\section{Accepted Parameters}{
\preformatted{list_group_memberships(
  GroupName = "string",
  NextToken = "string",
  MaxResults = 123,
  AwsAccountId = "string",
  Namespace = "string"
)
}
}

