% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.quicksight_operations.R
\name{list_user_groups}
\alias{list_user_groups}
\title{Lists the Amazon QuickSight groups that an Amazon QuickSight user is a member of}
\usage{
list_user_groups(UserName, AwsAccountId, Namespace, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{UserName}{[required] The Amazon QuickSight user name that you want to list group memberships for.}

\item{AwsAccountId}{[required] The AWS Account ID that the user is in. Currently, you use the ID for the AWS account that contains your Amazon QuickSight account.}

\item{Namespace}{[required] The namespace. Currently, you should set this to \code{default}.}

\item{NextToken}{A pagination token that can be used in a subsequent request.}

\item{MaxResults}{The maximum number of results to return from this request.}
}
\description{
Lists the Amazon QuickSight groups that an Amazon QuickSight user is a member of.
}
\details{
The permission resource is \code{arn:aws:quicksight:us-east-1:<aws-account-id>:user/default/<user-name> <aws-account-id>:user/default/<user-name> }.

The response is a one or more group objects.

\strong{CLI Sample:}

\code{aws quicksight list-user-groups -\-user-name=Pat -\-aws-account-id=111122223333 -\-namespace=default -\-region=us-east-1 }
}
\section{Accepted Parameters}{
\preformatted{list_user_groups(
  UserName = "string",
  AwsAccountId = "string",
  Namespace = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

