% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.quicksight_package.R
\docType{package}
\name{paws.quicksight-package}
\alias{paws.quicksight}
\alias{paws.quicksight-package}
\title{paws.quicksight: Amazon QuickSight}
\description{
Amazon QuickSight is a fully managed, serverless,
    cloud business intelligence service that makes it easy to extend data
    and insights to every user in your organization. This API interface
    reference contains documentation for a programming interface that you
    can use to manage Amazon QuickSight.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
