% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.quicksight_operations.R
\name{register_user}
\alias{register_user}
\title{Creates an Amazon QuickSight user, whose identity is associated with the AWS Identity and Access Management (IAM) identity or role specified in the request}
\usage{
register_user(
  IdentityType = "IAM"|"QUICKSIGHT",
  Email = "string",
  UserRole = "ADMIN"|"AUTHOR"|"READER"|"RESTRICTED_AUTHOR"|"RESTRICTED_READER",
  IamArn = "string",
  SessionName = "string",
  AwsAccountId = "string",
  Namespace = "string",
  UserName = "string"
)
}
\arguments{
\item{IdentityType}{[required] Amazon QuickSight supports several ways of managing the identity of users. This parameter accepts two values:
\itemize{
\item \code{IAM}: A user whose identity maps to an existing IAM user or role.
\item \code{QUICKSIGHT}: A user whose identity is owned and managed internally by Amazon QuickSight.
}}

\item{Email}{[required] The email address of the user that you want to register.}

\item{UserRole}{[required] The Amazon QuickSight role of the user. The user role can be one of the following:
\itemize{
\item \code{READER}: A user who has read-only access to dashboards.
\item \code{AUTHOR}: A user who can create data sources, data sets, analyses, and dashboards.
\item \code{ADMIN}: A user who is an author, who can also manage Amazon QuickSight settings.
}}

\item{IamArn}{The ARN of the IAM user or role that you are registering with Amazon QuickSight.}

\item{SessionName}{The name of the session with the assumed IAM role. By using this parameter, you can register multiple users with the same IAM role, provided that each has a different session name. For more information on assuming IAM roles, see \href{https://docs.aws.amazon.com/cli/latest/reference/sts/assume-role.html}{assume-role} in the \emph{AWS CLI Reference.}}

\item{AwsAccountId}{[required] The ID for the AWS account that the user is in. Currently, you use the ID for the AWS account that contains your Amazon QuickSight account.}

\item{Namespace}{[required] The namespace. Currently, you should set this to \code{default}.}

\item{UserName}{The Amazon QuickSight user name that you want to create for the user you are registering.}
}
\description{
Creates an Amazon QuickSight user, whose identity is associated with the AWS Identity and Access Management (IAM) identity or role specified in the request.
}
\details{
The permission resource is \code{arn:aws:quicksight:us-east-1:<aws-account-id>:user/default/<user-name> <aws-account-id>:user/default/<user-name> }.

The condition resource is the Amazon Resource Name (ARN) for the IAM user or role, and the session name.

The condition keys are \code{quicksight:IamArn} and \code{quicksight:SessionName}.

\strong{CLI Sample:}

\code{aws quicksight register-user -\-aws-account-id=111122223333 -\-namespace=default -\-email=pat@example.com -\-identity-type=IAM -\-user-role=AUTHOR -\-iam-arn=arn:aws:iam::111122223333:user/Pat }
}
\examples{

}
