% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.quicksight_operations.R
\name{update_group}
\alias{update_group}
\title{Changes a group description}
\usage{
update_group(
  GroupName = "string",
  Description = "string",
  AwsAccountId = "string",
  Namespace = "string"
)
}
\arguments{
\item{GroupName}{[required] The name of the group that you want to update.}

\item{Description}{The description for the group that you want to update.}

\item{AwsAccountId}{[required] The ID for the AWS account that the group is in. Currently, you use the ID for the AWS account that contains your Amazon QuickSight account.}

\item{Namespace}{[required] The namespace. Currently, you should set this to \code{default}.}
}
\description{
Changes a group description.
}
\details{
The permissions resource is \code{arn:aws:quicksight:us-east-1:<aws-account-id>:group/default/<group-name> <aws-account-id>:group/default/<group-name> }.

The response is a group object.

\strong{CLI Sample:}

\code{aws quicksight update-group --aws-account-id=111122223333 --namespace=default --group-name=Sales --description="Sales BI Dashboards" }
}
\examples{

}
